requirements.txt
# Core Dependencies
numpy>=1.24.0
gymnasium>=0.29.0
torch>=2.0.0
stable-baselines3>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
tensorboard>=2.13.0

# Optional: Advanced channel models
scipy>=1.10.0